requirements.txt
requests
beautifulsoup4
lxml
psycopg2-binary
python-dotenv
redis
google-search-results
